requirements.txt
django>=6.0.3
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-filter>=24.2
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
reportlab>=4.2.2
google-api-python-client>=2.169.0
google-auth>=2.40.0
google-auth-oauthlib>=1.2.2
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
